requirements.txt
pydantic>=2.0.0,<3.0.0
numpy==1.26.4
weaviate-client
langchain
langchain-openai
langchain-weaviate
datasets
ragas
openai

